Refactor GraphQL Start Crawl mutation into own file to make Search component default export
